@@ -17,25 +17,25 @@
   "devDependencies": {
     "@vue/compiler-sfc": "^3.4.21",
     "copy-webpack-plugin": "^12.0.2",
-    "css-loader": "^6.10.0",
+    "css-loader": "^6.11.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.6.0",
-    "sass-loader": "^14.1.1",
+    "html-webpack-plugin": "^5.6.3",
+    "sass-loader": "^16.0.4",
     "source-map-loader": "^5.0.0",
     "volar-service-vetur": "^0.0.34",
     "vue-loader": "^17.4.2",
-    "webpack": "^5.91.0",
+    "webpack": "^5.97.1",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4"
+    "webpack-dev-server": "^5.2.0"
   },
   "dependencies": {
     "bootstrap": "^5.3.3",
     "jquery": "^3.7.1",
     "js-cookie": "^3.0.5",
-    "sass": "^1.72.0",
+    "sass": "^1.83.1",
     "style-loader": "^3.3.4",
-    "vue": "^3.4.21",
-    "vue-cookies": "^1.8.3",
-    "vue-router": "^4.3.0"
+    "vue": "^3.5.13",
+    "vue-cookies": "^1.8.5",
+    "vue-router": "^4.5.0"
   }
 }
